web: file renaming, small improvements
