+++ internal/web/api/day.go
